backend/requirements.txt
fastapi>=0.115
uvicorn[standard]>=0.30
websockets>=12.0
python-dotenv>=1.0
python-multipart>=0.0.9
httpx>=0.27
pydantic>=2.11
pydantic-settings>=2.9
google-genai>=1.14
openai>=1.82
elevenlabs>=1.50
